test organized into entity packages and moved test around to appropriate class.
