+++ auth-service/src/main/java/com/petclinic/auth/User/UserIDLessDTO.java
@@ -1,5 +1,6 @@
 package com.petclinic.auth.User;
 
+import com.petclinic.auth.Config.PasswordStrengthCheck;
 import lombok.*;
 
 import javax.persistence.Column;
