+++ auth-service/src/test/java/com/petclinic/auth/AuthServiceApplicationTests.java
@@ -37,8 +37,6 @@ class AuthServiceApplicationTests {
 			ROLE_NAME = "role";
 	final Role role = new Role(0, ROLE_NAME);
 
-	private Validator validator;
-
 	final Set<Role> ROLES = new HashSet<Role>() {{
 		add(role);
 	}};
@@ -64,9 +62,6 @@ class AuthServiceApplicationTests {
 	@BeforeEach
 	void setup() {
 		roleRepo.deleteAllInBatch();
-		userRepo.deleteAllInBatch();
-		ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-		validator = factory.getValidator();
 	}
 
 	@Test
@@ -215,36 +210,6 @@ class AuthServiceApplicationTests {
 		assertEquals(ROLE_NAME, roleIDLessDTO.getName());
 		assertNull(roleIDLessDTO.getParent());
 	}
-
-	@Test
-	@DisplayName("Verify if the email is valid and succeed")
-	void verify_valid_email_success() {
-		User user = new User();
-		user.setUsername(USER);
-		user.setPassword(PASS);
-		user.setId(ID);
-		user.setEmail("testemail@gmail.com");
-		Set<ConstraintViolation<User>> violations = validator.validate(user);
-		assertTrue(violations.isEmpty());
-	}
-
-	@Test
-	@DisplayName("Verify if the email is valid and fail because missing @")
-	void detect_invalid_email_missing_at() {
-		User user = new User();
-		user.setUsername(USER);
-		user.setPassword(PASS);
-		user.setId(ID);
-		user.setEmail("testemailgmail.com");
-		Set<ConstraintViolation<User>> violations = validator.validate(user);
-		assertEquals(violations.size(), 1);
-
-		ConstraintViolation<User> violation = violations.iterator().next();
-		assertEquals("Email must be valid", violation.getMessage());
-		assertEquals("email", violation.getPropertyPath().toString());
-		assertEquals("testemailgmail.com", violation.getInvalidValue());
-	}
-
 	@Test
 	@DisplayName("Submit a completed signup form")
 	void submit_completed_signup_form() {
