+++ auth-service/src/test/java/com/petclinic/auth/User/AuthServiceUserServiceTests.java
@@ -1,4 +1,4 @@
-package com.petclinic.auth;
+package com.petclinic.auth.User;
 
 import com.petclinic.auth.Role.Role;
 import com.petclinic.auth.Role.RoleIDLessDTO;
@@ -49,10 +49,6 @@ public class AuthServiceUserServiceTests {
     @Autowired
     private UserService userService;
 
-
-    private UserController userController;
-    private final UserIDLessDTO ID_LESS_USER = new UserIDLessDTO();
-
     @BeforeEach
     void setup() {
         userRepo.deleteAllInBatch();
@@ -89,7 +85,6 @@ public class AuthServiceUserServiceTests {
     @DisplayName("Reset password, passed wrong ID")
     void test_user_password_reset_ID() {
 
-        final String CHANGE = "change";
         final User u = new User(USER, PASS, EMAIL);
         userRepo.save(u);
 
@@ -101,7 +96,7 @@ public class AuthServiceUserServiceTests {
     void create_new_user_with_same_email() {
         UserIDLessDTO userIDLessDTO = new UserIDLessDTO(USER, PASS, EMAIL);
         User userMap = userMapper.idLessDTOToModel(userIDLessDTO);
-        User saved = userRepo.save(userMap);
+        userRepo.save(userMap);
         assertThrows(DuplicateKeyException.class, () -> userService.createUser(userIDLessDTO));
     }
 }
